requirements.txt
pillow>=9.0.0
pandas>=1.3.0
pytesseract>=0.3.8
